install: Move options logic into options package
