+++ options/install/flags.go
@@ -14,7 +14,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package options
+package install
 
 import (
 	"github.com/spf13/cobra"
